Create MatrixConversion
Takes a given matrix in any numerical format and converts it to one of either floating point or integer valued elements, under the assumption that the elements within the matrix are NOT strings or any other incompatible data type. This is necessary when performing computations such as placing the matrix into triangular form, Gaussian elimination, and etc.